requirements.txt
pyrogram
tgcrypto
gallery-dl
Pillow
Flask
motor
aiohttp
